Make getCurrentTimeBackup() generate equal values for a backup
MysqlBackupTool.getCurrentTimeBackup() is called in 2 different places
to get the backup time, and this drift causes it to attempt to write the
files it saved to a timestamped directory different than the timestamped
directory it actually created. Fix this by converting it to store its
results in an instance variable.